Modified behavour of narrative updates.
In reference with the category select, if the user changes their mind, they can simply focus away from the select drop down in order to remove it from view in order to select another narrative.

The interface will also be more interactive by blinking the row after a successful operation.